Create tests/test_blobstore.py
